# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/user/create-user": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Create User Without Password",
        "operationId": "create_user_without_password_api_v1_user_create_user_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateWithoutPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCreateWithoutPasswordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/user/set-password": {
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "Set User Password",
        "operationId": "set_user_password_api_v1_user_set_password_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSetPasswordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/me": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get Me",
        "operationId": "get_me_api_v1_user_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update Me",
        "operationId": "update_me_api_v1_user_me_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/department/": {
      "post": {
        "tags": [
          "department"
        ],
        "summary": "Create Department",
        "operationId": "create_department_api_v1_department__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/department/{dep_id}": {
      "get": {
        "tags": [
          "department"
        ],
        "summary": "Get Department",
        "operationId": "get_department_api_v1_department__dep_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dep_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dep Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "department"
        ],
        "summary": "Update Department",
        "operationId": "update_department_api_v1_department__dep_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dep_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dep Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "department"
        ],
        "summary": "Delete Department",
        "operationId": "delete_department_api_v1_department__dep_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dep_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dep Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/department/set_head": {
      "post": {
        "tags": [
          "department"
        ],
        "summary": "Set Department Head",
        "operationId": "set_department_head_api_v1_department_set_head_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DepartmentSetHeadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DepartmentCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "parent_department_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Department Id"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "company_id"
        ],
        "title": "DepartmentCreateRequest"
      },
      "DepartmentDeleteResponse": {
        "properties": {
          "status": {
            "type": "boolean",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "DepartmentDeleteResponse"
      },
      "DepartmentResponse": {
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/DepartmentSchema"
          }
        },
        "type": "object",
        "required": [
          "payload"
        ],
        "title": "DepartmentResponse"
      },
      "DepartmentSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "parent_department_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Department Id"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "head_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Head Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "company_id",
          "id",
          "path"
        ],
        "title": "DepartmentSchema"
      },
      "DepartmentSetHeadRequest": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "department_id": {
            "type": "integer",
            "title": "Department Id"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "department_id"
        ],
        "title": "DepartmentSetHeadRequest"
      },
      "DepartmentUpdateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "DepartmentUpdateRequest"
      },
      "GetMeResponse": {
        "properties": {
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "department_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "email",
          "company_id",
          "role",
          "id"
        ],
        "title": "GetMeResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserCreateWithoutPasswordRequest": {
        "properties": {
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "email"
        ],
        "title": "UserCreateWithoutPasswordRequest"
      },
      "UserCreateWithoutPasswordResponse": {
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/UserSchema"
          }
        },
        "type": "object",
        "required": [
          "payload"
        ],
        "title": "UserCreateWithoutPasswordResponse"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "admin",
          "company_admin",
          "user"
        ],
        "title": "UserRole"
      },
      "UserSchema": {
        "properties": {
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "company_id": {
            "type": "integer",
            "title": "Company Id"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "department_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "email",
          "company_id",
          "role",
          "id"
        ],
        "title": "UserSchema"
      },
      "UserSetPasswordRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserSetPasswordRequest"
      },
      "UserSetPasswordResponse": {
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/UserSchema"
          }
        },
        "type": "object",
        "required": [
          "payload"
        ],
        "title": "UserSetPasswordResponse"
      },
      "UserUpdateRequest": {
        "properties": {
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "last_name"
        ],
        "title": "UserUpdateRequest"
      },
      "UserUpdateResponse": {
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/UserSchema"
          }
        },
        "type": "object",
        "required": [
          "payload"
        ],
        "title": "UserUpdateResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}